more restrict communication pattern

+++ src/net/vicp/lylab/core/CoreDef.java
@@ -134,7 +134,7 @@ public class CoreDef extends NonCloneableBaseObject {
 	public static final int DEFAULT_SOCKET_WRITE_TTIMEOUT = (int) (TWO * MINUTE);
 
 	public static final int DEFAULT_LYCACHE_CONTAINER_SIZE = SIXTEEN;
-	public static final long DEFAULT_LYCACHE_EXPIRE_TIME = FIFTEEN * TWO * MINUTE;
+	public static final int DEFAULT_LYCACHE_EXPIRE_TIME = (int) (FIFTEEN * TWO * MINUTE);
 	public static final long DEFAULT_LYCACHE_MEMORY_LIMITATION = ONE * GB;
 	public static final double DEFAULT_LYCACHE_THRESHOLD = DOUBLE * EIGHT / TEN;
